archive/builder_pack/builder/export/MusicAdvisor/HitCheck/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "musicadvisor-hitcheck"
version = "1.0.0"
description = "HitCheck Layer for Music Advisor v1.1.1 — kNN benchmarking, drift mapping, echo lineage, and HCI_v1p projection"
authors = [{name="Music Advisor Team"}]
readme = "README_HitCheck.md"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.23",
  "scikit-learn>=1.1",
  "pyyaml>=6.0.1"
]

[tool.setuptools.package-data]
"hitcheck" = ["../schemas/*.json", "config.yaml"]
